requirements.txt
agate==1.9.1
altair==5.4.1
annotated-types==0.7.0
appnope==0.1.4
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
babel==2.16.0
backports-tarfile==1.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
daff==1.3.46
dbt-adapters==1.4.1
dbt-common==1.8.0
dbt-core==1.8.6
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
dlt==1.0.0
dotenv-python==0.0.1
executing==2.1.0
filelock==3.16.0
fsspec==2024.9.0
git-filter-repo==2.45.0
gitdb==4.0.11
gitpython==3.1.43
giturlparse==0.12.0
hexbytes==1.2.1
humanize==4.10.0
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.0.2
jedi==0.19.1
jinja2==3.1.4
jsonpath-ng==1.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.2
jupyter-core==5.7.2
keyring==25.3.0
leather==0.4.0
logbook==1.5.3
makefun==1.15.4
markdown-it-py==3.0.0
markupsafe==2.1.5
mashumaro==3.13.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
mysql-connector-python==9.0.0
narwhals==1.9.2
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
ordered-set==4.1.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.1
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.2
ply==3.11
prompt-toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.1
pydantic-core==2.23.3
pydeck==0.9.1
pygments==2.18.0
pyjwt==2.9.0
pyopenssl==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requirements-parser==0.11.0
rich==13.9.2
rpds-py==0.20.0
semver==3.0.2
setuptools==75.1.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.39.0
tenacity==9.0.0
text-unidecode==1.3
time-machine==2.15.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
types-setuptools==75.1.0.20240917
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wheel==0.44.0
zipp==3.20.1
